Updated FunderToChanneler to have UpdateFriend.
